// src/payload-types.ts
export interface Header {
  id: string;
  navLinks?:
    | {
        label: string;
        link?: {
          type?: ('internal' | 'external') | null;
          reference?: (string | null) | Page;
          url?: string | null;
        };
        newTab?: boolean | null;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}